fix negative coordinates in input

+++ src/input.rs
@@ -5,14 +5,15 @@ type C<'a> = wasmi::Caller<'a, State>;
 
 pub(crate) fn read_left(mut caller: C) -> u32 {
     let state = caller.data_mut();
+    // TODO: cache input in state
     let input = match state.device.read_input() {
         Some(InputState {
             left: Some(left), ..
         }) => left,
         _ => return 0,
     };
-    let x = input.x as u32;
-    let y = input.y as u32;
+    let x = input.x as u16 as u32;
+    let y = input.y as u16 as u32;
     x << 16 | y
 }
